Apply feedback for R0 astromech
