+++ Assets/Scripts/Model/Content/SecondEdition/Pilots/Resistance/T70XWing/JaycrisTubbsEoD.cs
@@ -1,9 +1,5 @@
-using System;
 using System.Collections.Generic;
 using Content;
-using Movement;
-using Ship;
-using Tokens;
 using Upgrade;
 using UpgradesList.SecondEdition;
 
@@ -39,83 +35,6 @@ namespace Ship.SecondEdition.T70XWing
             MustHaveUpgrades.Add(typeof(ROAstromech));
 
             PilotNameCanonical = "jaycristubbs-evacuationofdqar";
-        }
-    }
-}
-
-namespace UpgradesList.SecondEdition
-{
-    public class ROAstromech : GenericUpgrade
-    {
-        public ROAstromech() : base()
-        {
-            UpgradeInfo = new UpgradeCardInfo(
-                "RO Astromech",
-                UpgradeType.Astromech,
-                abilityType: typeof(Abilities.SecondEdition.ROAstromechAbility),
-                legalityInfo: new() { Legality.StandardLegal, Legality.ExtendedLegal, Legality.XWA }
-            );
-            IsHidden = true;
-        }
-    }
-}
-
-namespace Abilities.SecondEdition
-{
-    // After you reveal a basic maneuver, you may reduce its difficulty.
-    // If you do, after the Check Difficulty step, gain 1 strain token.
-    public class ROAstromechAbility : GenericAbility
-    {
-        public override void ActivateAbility()
-        {
-            HostShip.OnManeuverIsRevealed += CheckManeuverForAbility;
-        }
-
-        public override void DeactivateAbility()
-        {
-            HostShip.OnManeuverIsRevealed -= CheckManeuverForAbility;
-        }
-
-        private void CheckManeuverForAbility(GenericShip ship)
-        {
-            if (HostShip.AssignedManeuver.IsBasicManeuver
-                && (HostShip.AssignedManeuver.ColorComplexity == Movement.MovementComplexity.Normal || HostShip.AssignedManeuver.ColorComplexity == Movement.MovementComplexity.Complex)
-            )
-            {
-                RegisterAbilityTrigger(TriggerTypes.OnManeuverIsRevealed, RegisterROAstromechAbility);
-            }
-        }
-
-        private void RegisterROAstromechAbility(object sender, EventArgs e)
-        {
-            AskToUseAbility
-            (
-                HostUpgrade.UpgradeInfo.Name,
-                DoesAiUse,
-                UseROAstromechAbility,
-                descriptionLong: "Do you want to reduce the difficulty of your maneuver? If you do, gain 1 Strain token after the check difficulty step.",
-                imageHolder: HostUpgrade
-            );
-        }
-
-        private void UseROAstromechAbility(object sender, EventArgs e)
-        {
-            HostShip.AssignedManeuver.ColorComplexity = GenericMovement.ReduceComplexity(HostShip.AssignedManeuver.ColorComplexity);
-            Triggers.FinishTrigger();
-            HostShip.OnMovementFinish += GainStrain;
-        }
-
-        private void GainStrain(GenericShip ship)
-        {
-            HostShip.OnMovementFinish -= GainStrain;
-            HostShip.Tokens.AssignToken(new StrainToken(HostShip),()=>{});
-        }
-
-        private bool DoesAiUse()
-        {
-            return HostShip.Tokens.HasToken<StressToken>()
-                || HostShip.Tokens.HasToken<StrainToken>()
-                || HostShip.Tokens.HasToken<DepleteToken>(); // This could be improved to only swap deplete for strain when it expects to attack this turn.
         }
     }
 }
