+++ Assets/Scripts/Model/Content/SecondEdition/Upgrades/Astromech/R0.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f42c92fd7adfff74d81efa816ec64e6b
